% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bhmtrains}
\alias{bhmtrains}
\title{Birmingham Trains, Dec 2016-Feb 2017.}
\format{A data frame with 83710 rows and 16 variables:
\describe{
  \item{ServiceId}{Unique train identifier}
  \item{Status}{Train status: A=Active, C=Cancelled, R=Reinstated}
  \item{TOC}{Train operating company}
  \item{TrainCategory}{Express Passenger or Ordinary Passenger}
  \item{PowerType}{DMU=Diesel Multiple Unit, EMU=Electrical Multiple Unit,
  HST=High Speed Train}
  \item{SchedSpeedMPH}{Scheduled maximum speed (in miles per hour)}
  \item{Origin}{3-letter code denoting the scheduled origin of the train}
  \item{OriginGbttDeparture}{Scheduled departure time in the Great Britain
  Train Timetable (GBTT) from the origin station}
  \item{OriginActualDeparture}{Actual departure time from the origin station}
  \item{GbttArrival}{Scheduled arrival time in Birmingham in the GBTT}
  \item{ActualArrival}{Actual arrival time in Birmingham in the GBTT}
  \item{GbttDeparture}{Scheduled departture time from Birmingham in the GBTT}
  \item{ActualDeparture}{Actual departure time from Birmingham in the GBTT}
  \item{Destination}{3-letter code denoting the scheduled destination of the train}
  \item{DestinationGbttArrival}{Scheduled arrival time in the GBTT at the
  destination station}
  \item{DestinationActualArrival}{Actual arrival time at the destination station}
}}
\source{
\url{http://www.recenttraintimes.co.uk/}
}
\usage{
bhmtrains
}
\description{
A dataset containing all of the trains that either
originated at, passed through or terminated at
Birmingham New Street railway station in the UK
between 1st December 2016 and 28th February 2017
inclusive.
}
\keyword{datasets}
