% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Template.R
\docType{class}
\name{ClassName}
\alias{ClassName}
\title{A class that defines a calculation.}
\format{\code{\link{R6Class}} object.}
\usage{
ClassName
}
\value{
Object of \code{\link{R6Class}} with properties and methods that define a single pivot table calculation.
}
\description{
The PivotCalculation class defines one calculation in a pivot table.
}
\section{Fields}{

\describe{
\item{\code{parentPivot}}{Owning pivot table.}

\item{\code{calculationName}}{Calculation unique name.  Recommendation:  Do not have spaces in this name.}
}}

\section{Methods}{

\describe{
  \item{Documentation}{For more complete explanations and examples please see the extensive vignettes supplied with this package.}
  \item{\code{new(...)}}{Create a new pivot calculation, specifying the field values documented above.}

  \item{\code{asList()}}{Get a list representation of this calculation.}
}
}

\examples{
This class should only be created by the pivot table.
It is not intended to be created outside of the pivot table.
}
\keyword{calculation}
