% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleParsing.R
\name{cleanCssValue}
\alias{cleanCssValue}
\title{Cleans up a CSS attribute value.}
\usage{
cleanCssValue(cssValue)
}
\arguments{
\item{cssValue}{The value to cleanup.}
}
\value{
The cleaned value.
}
\description{
\code{cleanCssValue} is a utility function that performs some basic cleanup
on CSS attribute values.  Leading and trailing whitespace is removed.  The
CSS values "initial" and "inherit" are blocked.  The function is vectorised
so can be used with arrays.
}
