% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cleanOutlineArg}
\alias{cleanOutlineArg}
\title{Clean the arguments specified for an outline group}
\usage{
cleanOutlineArg(
  pvt,
  outline = NULL,
  defaultCaption = "{value}",
  defaultIsEmpty = TRUE
)
}
\arguments{
\item{pvt}{The pivot table.}

\item{outline}{Either a logical value (TRUE to use the default
outline settings) or a list specifying outline settings.}

\item{defaultCaption}{The default caption of the outline group.}

\item{defaultIsEmpty}{Specify whether the outline group is
empty or contains a value (typically a sub-total)}
}
\value{
A listed containing checked/cleaned outline group
settings.
}
\description{
\code{cleanOutlineArg} checks values and provides defaults.
}
