% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{containsText}
\alias{containsText}
\title{Check whether a text value is present in another text value.}
\usage{
containsText(textToSearch, textToFind)
}
\arguments{
\item{textToSearch}{The value to be searched.}

\item{textToFind}{The value to find.}
}
\value{
TRUE if the textToFind value is found.
}
\description{
\code{containsText} is a utility function returns TRUE if one text value is
present in another.  Case sensitive.  If textToSearch is a vector, returns
TRUE if any element contains textToFind.
}
