% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertor.R
\name{convertPvtStyleToBasicStyle}
\alias{convertPvtStyleToBasicStyle}
\title{Convert a pivot table style to a basictabler style.}
\usage{
convertPvtStyleToBasicStyle(btbl = NULL, pvtStyle = NULL)
}
\arguments{
\item{btbl}{The basic table that will own the new style.}

\item{pvtStyle}{The pivot style to convert.}
}
\value{
a basictabler style.
}
\description{
\code{convertPvtStyleToBasicStyle} is a utility function that converts a pivot
table style to a basictabler style from the basictabler package.
}
