% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertor.R
\name{convertPvtTblToBasicTbl}
\alias{convertPvtTblToBasicTbl}
\title{Convert a pivot table to a basic table.}
\usage{
convertPvtTblToBasicTbl(pvt = NULL, exportOptions = NULL)
}
\arguments{
\item{pvt}{The pivot table to convert.}

\item{exportOptions}{Options specifying how values are exported.}
}
\value{
a basictabler table.
}
\description{
\code{convertPvtTblToBasicTbl} is a utility function that converts a pivot
table to a basic table from the basictabler package.
}
