% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expandHex}
\alias{expandHex}
\title{Expand hex.}
\usage{
expandHex(hex)
}
\arguments{
\item{hex}{The hex value to be expanded.}
}
\value{
The expanded hex value.
}
\description{
\code{expandHex} is a utility function that expands 3-digit hex to 6-digit.
}
