% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{exportValueAs}
\alias{exportValueAs}
\title{Replace the current value with a placeholder during export.}
\usage{
exportValueAs(
  rawValue,
  formattedValue,
  exportOptions,
  blankValue = character(0)
)
}
\arguments{
\item{rawValue}{The raw value to check.}

\item{formattedValue}{The formatted value to be exported.}

\item{exportOptions}{A list of options controlling export behaviour.}

\item{blankValue}{The 'placeholder' value to be exported when skipping the value.}
}
\value{
Either the original value or a placeholder value.
}
\description{
\code{exportValueAs} is a utility function that returns either the
original value or a replacement placeholder value for export.
}
