% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotThemes.R
\name{getCompactTheme}
\alias{getCompactTheme}
\title{Get the compact theme for styling a pivot table.}
\usage{
getCompactTheme(parentPivot, themeName = "compact")
}
\arguments{
\item{parentPivot}{Owning pivot table.}

\item{themeName}{The name to use as the new theme name.}
}
\value{
A PivotStyles object.
}
\description{
Get the compact theme for styling a pivot table.
}
