% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotThemes.R
\name{getDefaultTheme}
\alias{getDefaultTheme}
\title{Get the default theme for styling a pivot table.}
\usage{
getDefaultTheme(parentPivot, themeName = "default")
}
\arguments{
\item{parentPivot}{Owning pivot table.}

\item{themeName}{The name to use as the new theme name.}
}
\value{
A PivotStyles object.
}
\description{
Get the default theme for styling a pivot table.
}
