% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotThemes.R
\name{getLargePlainTheme}
\alias{getLargePlainTheme}
\title{Get the large plain theme for styling a pivot table.}
\usage{
getLargePlainTheme(parentPivot, themeName = "largeplain")
}
\arguments{
\item{parentPivot}{Owning pivot table.}

\item{themeName}{The name to use as the new theme name.}
}
\value{
A PivotStyles object.
}
\description{
Get the large plain theme for styling a pivot table.
}
