% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleParsing.R
\name{getNextPosition}
\alias{getNextPosition}
\title{Find the first value in an array that is larger than the specified value.}
\usage{
getNextPosition(positions, afterPosition)
}
\arguments{
\item{positions}{An ordered numeric vector.}

\item{afterPosition}{The value to start searching after.}
}
\value{
The first value in the array larger than afterPosition.
}
\description{
\code{getNextPosition} is a utility function that helps when parsing strings
that contain delimiters.
}
