% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertor.R
\name{getPvtStyleDeclarations}
\alias{getPvtStyleDeclarations}
\title{Get pivot table style declarations from a pivot table style.}
\usage{
getPvtStyleDeclarations(pvtStyle = NULL)
}
\arguments{
\item{pvtStyle}{The pivot table style to read.}
}
\value{
a list of style declarations.
}
\description{
\code{getPvtStyleDeclarations} is a utility function that reads the styles
from a pivot table style.
}
