% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PivotThemes.R
\name{getSimpleColoredTheme}
\alias{getSimpleColoredTheme}
\title{Get a simple coloured theme.}
\usage{
getSimpleColoredTheme(parentPivot, themeName = "coloredTheme", colors,
  fontName)
}
\arguments{
\item{parentPivot}{Owning pivot table.}

\item{themeName}{The name to use as the new theme name.}

\item{colors}{The set of colours to use when generating the theme (see the Styling vignette for details).}

\item{fontName}{The name of the font to use, or a comma separated list (for font-fall-back).}
}
\value{
A PivotStyles object.
}
\description{
Get a simple coloured theme that can be used to style a pivot table into a custom colour scheme.
}
