% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotThemes.R
\name{getStandardTableTheme}
\alias{getStandardTableTheme}
\title{Get the a theme for styling to a pivot table that looks
more like a standard table (i.e. no row column headings).}
\usage{
getStandardTableTheme(parentPivot, themeName = "standardtable")
}
\arguments{
\item{parentPivot}{Owning pivot table.}

\item{themeName}{The name to use as the new theme name.}
}
\value{
A PivotStyles object.
}
\description{
Get the a theme for styling to a pivot table that looks
more like a standard table (i.e. no row column headings).
}
