% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotThemes.R
\name{getTheme}
\alias{getTheme}
\title{Get a built-in theme for styling a pivot table.}
\usage{
getTheme(parentPivot, themeName = NULL)
}
\arguments{
\item{parentPivot}{Owning pivot table.}

\item{themeName}{The name of the theme to retrieve.}
}
\value{
A PivotStyles object.
}
\description{
\code{getTheme} returns the specified theme.
}
