% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleParsing.R
\name{getXlBorderStyleFromCssBorder}
\alias{getXlBorderStyleFromCssBorder}
\title{Convert CSS border values to those used by the openxlsx package.}
\usage{
getXlBorderStyleFromCssBorder(border)
}
\arguments{
\item{border}{A list containing elements width and style.}
}
\value{
An openxlsx border style.
}
\description{
\code{getXlBorderStyleFromCssBorder} takes border parameters expressed as a
list (containing elements: width and style) and returns a border style that
is compatible with the openxlsx package.
}
