% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isNumericValue}
\alias{isNumericValue}
\title{Check whether a numeric value is present.}
\usage{
isNumericValue(value)
}
\arguments{
\item{value}{The value to check.}
}
\value{
TRUE if a numeric value is present.
}
\description{
\code{isNumericValue} is a utility function returns TRUE only when a numeric value
is present.  NULL, NA, numeric(0) and integer(0) all return FALSE.
}
