% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isTextValue}
\alias{isTextValue}
\title{Check whether a text value is present.}
\usage{
isTextValue(value)
}
\arguments{
\item{value}{The value to check.}
}
\value{
TRUE if a non-blank text value is present.
}
\description{
\code{isTextValue} is a utility function returns TRUE only when a text value
is present.  NULL, NA, character(0) and "" all return FALSE.
}
