% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{oneToNULL}
\alias{oneToNULL}
\title{Convert a value of 1 to a NULL value.}
\usage{
oneToNULL(value, convertOneToNULL)
}
\arguments{
\item{value}{The value to check.}

\item{convertOneToNULL}{TRUE to convert 1 to NULL.}
}
\value{
NULL if value==1, otherwise value.
}
\description{
\code{oneToNull} is a utility function that returns NULL when a value of
0 or 1 is passed to it, otherwise it returns the original value.
}
