% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleParsing.R
\name{parseColor}
\alias{parseColor}
\title{Convert a CSS colour into a hex based colour code.}
\usage{
parseColor(color)
}
\arguments{
\item{color}{The colour to convert.}
}
\value{
The colour as a hex code, e.g. #FF00A0.
}
\description{
\code{parseColor} converts a colour value specified in CSS to a hex based
colour code.  Example supported input values/formats/named colours are:
#0080FF, rgb(0, 128, 255), rgba(0, 128, 255, 0.5) and red, green, etc.
}
