% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleParsing.R
\name{parseCssSizeToPt}
\alias{parseCssSizeToPt}
\title{Convert a CSS size value into points.}
\usage{
parseCssSizeToPt(size)
}
\arguments{
\item{size}{A size specified in common CSS units.}
}
\value{
The size converted to points.
}
\description{
\code{parseCssSizeToPt} will take a CSS style and convert it to points.
Supported input size units are in, cm, mm, pt, pc, px, em, %.  The following
are converted exactly:  in, cm, mm, pt, pc: using 1in = 2.54cm = 25.4mm =
72pt = 6pc.  The following are converted approximately:  px, em, %: using
approx 1em=16px=12pt and 100%=1em, so approx 25.4mm = 96px
}
