% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleParsing.R
\name{parseCssSizeToPx}
\alias{parseCssSizeToPx}
\title{Convert a CSS size value into pixels}
\usage{
parseCssSizeToPx(size)
}
\arguments{
\item{size}{A size specified in common CSS units.}
}
\value{
The size converted to pixels.
}
\description{
\code{parseCssSizeToPx} will take a CSS style and convert it to pixels
Supported input size units are in, cm, mm, pt, pc, px, em, %.  The following
are converted exactly:  in, cm, mm, pt, pc: using 1in = 2.54cm = 25.4mm =
72pt = 6pc.  The following are converted approximately:  px, em, %: using
approx 1em=16px=12pt and 100%=1em, so approx 25.4mm = 96px
}
