% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleParsing.R
\name{parseCssString}
\alias{parseCssString}
\title{Split a CSS attribute value into a vector/array.}
\usage{
parseCssString(text, separator = ",", removeEmptyString = TRUE)
}
\arguments{
\item{text}{The text to split.}

\item{separator}{The field separator, default comma.}

\item{removeEmptyString}{TRUE to not return empty string / whitespace values.}
}
\value{
An R vector containing the values from text split up.
}
\description{
\code{parseCssString} is a utility function that splits a string into a
vector/array.  The function pays attention to text qualifiers (single and
double quotes) so won't split if the delimiter occurs inside a value.
}
