% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleParsing.R
\name{parseXlBorder}
\alias{parseXlBorder}
\title{Parse an xl-border value.}
\usage{
parseXlBorder(text)
}
\arguments{
\item{text}{The border declaration to parse.}
}
\value{
A list containing two elements: style and color.
}
\description{
\code{parseXlBorder} parses the combined xl border declarations (i.e.
xl-border, xl-border-left, xl-border-right, xl-border-top, xl-border-bottom)
and returns a list containing style and color as separate elements.
}
