% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PivotBatch.R
\docType{class}
\name{PivotBatch}
\alias{PivotBatch}
\title{A class that represents a batch calculation.}
\format{\code{\link{R6Class}} object.}
\usage{
PivotBatch
}
\value{
Object of \code{\link{R6Class}} with properties and methods that help
  to (do xyz).
}
\description{
The PivotBatch class represents one combination of variables and calculations
that are needed when calculating the values of cells in a pivot table.
}
\section{Fields}{

\describe{
\item{\code{parentPivot}}{Owning pivot table.}
}}

\section{Methods}{

\describe{
  \item{Documentation}{For more complete explanations and examples please see
  the extensive vignettes supplied with this package.}
  \item{\code{new(...)}}{Create a new Pivot Batch.}

  \item{\code{exampleMethod(df, dataName)}}{Example to be removed later.}
}
}

\examples{
# This class should only be created by the pivot table.
# It is not intended to be created outside of the pivot table.
}
\keyword{datasets}
