% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PivotHtmlRenderer.R
\docType{class}
\name{PivotHtmlRenderer}
\alias{PivotHtmlRenderer}
\title{A class that renders a pivot table in HTML.}
\format{\code{\link{R6Class}} object.}
\usage{
PivotHtmlRenderer
}
\value{
Object of \code{\link{R6Class}} with properties and methods that
  render to HTML.
}
\description{
The PivotHtmlRenderer class creates a HTML representation of a pivot table.
}
\section{Fields}{

\describe{
\item{\code{parentPivot}}{Owning pivot table.}
}}

\section{Methods}{

\describe{
  \item{Documentation}{For more complete explanations and examples please see
  the extensive vignettes supplied with this package.}
  \item{\code{new(...)}}{Create a new pivot table renderer, specifying the
  field value documented above.}

  \item{\code{clearIsRenderedFlags()}}{Clear the IsRendered flags that exist
  on the PivotDataGroup class.}
  \item{\code{getTableHtml(styleNamePrefix=NULL, includeHeaderValues=FALSE,
  includeRCFilters=FALSE, includeCalculationFilters=FALSE,
  includeCalculationNames=FALSE, includeRawValue=FALSE,
  includeTotalInfo=FALSE)}}{Get a HTML representation of the pivot table,
  optionally including additional detail for debugging purposes.}
}
}

\examples{
# This class should only be created by the pivot table.
# It is not intended to be created outside of the pivot table.
}
\keyword{datasets}
