% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivottabler.R
\name{pivottabler}
\alias{pivottabler}
\title{Render a pivot table as a HTML widget.}
\usage{
pivottabler(
  pt,
  width = NULL,
  height = NULL,
  styleNamePrefix = NULL,
  includeRCFilters = FALSE,
  includeCalculationFilters = FALSE,
  includeWorkingData = FALSE,
  includeEvaluationFilters = FALSE,
  includeCalculationNames = FALSE,
  includeRawValue = FALSE
)
}
\arguments{
\item{pt}{The pivot table to render.}

\item{width}{The target width.}

\item{height}{The target height.}

\item{styleNamePrefix}{A text prefix to be prepended to the CSS declarations
(to ensure uniqueness).}

\item{includeRCFilters}{Show/hide filter detail for debugging.}

\item{includeCalculationFilters}{Show/hide filter detail for debugging.}

\item{includeWorkingData}{Show/hide working data detail for debugging.}

\item{includeEvaluationFilters}{Show/hide filter detail for debugging.}

\item{includeCalculationNames}{Show/hide filter detail for debugging.}

\item{includeRawValue}{Show/hide filter detail for debugging.}
}
\value{
A HTML widget.
}
\description{
The \code{pivottabler} function is primarily intended for use with Shiny web
applications.
}
\examples{
# See the Shiny vignette in this package for examples.
}
