% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivottabler.R
\name{pivottablerOutput}
\alias{pivottablerOutput}
\title{Standard function for Shiny scaffolding.}
\usage{
pivottablerOutput(outputId, width = "100\%", height = "100\%")
}
\description{
Standard function for Shiny scaffolding.
}
