% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{processIdentifiers}
\alias{processIdentifiers}
\title{Handle identifiers that may be illegal (e.g. containing spaces).}
\usage{
processIdentifiers(identifiers)
}
\arguments{
\item{identifiers}{The identifiers that may be illegal.}
}
\value{
The identifiers wrapped in backticks (if illegal) or unchanged.
}
\description{
\code{processIdentifiers} is a utility function that wraps illegal
identifiers in backticks.
}
