% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pvtperfresults}
\alias{pvtperfresults}
\title{Performance Comparison Results}
\format{A data frame with 216 rows and  11 variables:
\describe{
  \item{id}{A unique identifier for this test case.}
  \item{evaluationMode}{The pivot table evaluation mode used for this test case.}
  \item{rowCount}{The number of rows in the data frame used for this test case.}
  \item{cellCount}{The number of cells in the pivot table used for this test case.}
  \item{argumentCheckMode}{The pivot table argument check mode used this test case.}
  \item{processingLibrary}{The pivot table processing library used this test case.}
  \item{description}{A description of this test case.}
  \item{completed}{A logical value indicating whether this test case completed.}
  \item{userTime}{The user time for this test case.}
  \item{systemTime}{The system time for this test case.}
  \item{elapsedTime}{The elapsed time for this test case.}
}}
\usage{
pvtperfresults
}
\description{
A reference dataset containing the full results of an example performance
comparison for different pivot table test cases.
}
\keyword{datasets}
