% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pvtperftests}
\alias{pvtperftests}
\title{Performance Comparison Results}
\format{A data frame with 26 rows and  18 variables:
\describe{
  \item{id}{A unique identifier for this test scenario.}
  \item{evaluationMode}{The pivot table evaluation mode used for this test scenario.}
  \item{rowCount}{The number of rows in the data frame used for this test scenario.}
  \item{cellCount}{The number of cells in the pivot table used for this test scenario.}
  \item{argumentCheckMode}{The pivot table argument check mode used this test scenario.}
  \item{processingLibrary}{The pivot table processing library used this test scenario.}
  \item{description}{A description of this test scenario.}
  \item{userTimeAvg}{The average user time for this test scenario.}
  \item{systemTimeAvg}{The average system time for this test scenario.}
  \item{elapsedTimeAvg}{The average elapsed time for this test scenario.}
  \item{userTimeMin}{The minimum user time for this test scenario.}
  \item{userTimeMax}{The maximum user time for this test scenario.}
  \item{systemTimeMin}{The minimum system time for this test scenario.}
  \item{systemTimeMax}{The maximum system time for this test scenario.}
  \item{elapsedTimeMin}{The minimum elapsed time for this test scenario.}
  \item{elapsedTimeMax}{The maximum elapsed time for this test scenario.}
  \item{testName}{A short name for this test scenario.}
  \item{testIndex}{An index for this test scenario.}
}}
\usage{
pvtperftests
}
\description{
A reference dataset containing an example performance comparison for
different pivot table test scenarios.
}
\keyword{datasets}
