% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicTable.R
\name{renderBasicTable}
\alias{renderBasicTable}
\title{Output a table into a package vignette.}
\usage{
renderBasicTable(matrix = NULL, stylePrefix = NULL, rowNamesAsHeader = FALSE)
}
\arguments{
\item{matrix}{Tabular data to render.}

\item{stylePrefix}{Text prefix for CSS style declarations.}

\item{rowNamesAsHeader}{Include row names in output.}
}
\value{
A basic table rendered as a HTML widget.
}
\description{
\code{renderBasicTable} is utility function that renders a basic table into a
package vignette.  This function is primarily intended for internal use by
the pivottabler package.
}
\examples{
renderBasicTable(matrix(c(1:12), nrow=3))
}
