% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivottabler.R
\name{renderPivottabler}
\alias{renderPivottabler}
\title{Standard function for Shiny scaffolding.}
\usage{
renderPivottabler(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Standard function for Shiny scaffolding.
}
