% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{skipExportingValue}
\alias{skipExportingValue}
\title{Should the current value be skipped during export?}
\usage{
skipExportingValue(rawValue, exportOptions)
}
\arguments{
\item{rawValue}{The value to check.}

\item{exportOptions}{A list of options controlling export behaviour.}
}
\value{
TRUE or FALSE indicating whether the current value should be
skipped.
}
\description{
\code{skipExportingValue} is a utility function that returns true if
the current value should be skipped when exporting.
}
