% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trainstations}
\alias{trainstations}
\title{Train Stations}
\format{
A data frame with 2568 rows and  7 variables:
\describe{
  \item{CrsCode}{3-letter code for the station}
  \item{StationName}{The name of the station}
  \item{OsEasting}{The UK Ordnance Survey Easting coordinate for the station}
  \item{OsNorthing}{The UK Ordnance Survey Northing coordinate for the
  station}
  \item{GridReference}{Grid reference for the station}
  \item{Latitude}{Latitude of the station location}
  \item{Longitude}{Longitude of the station location}
}
}
\source{
\url{https://www.recenttraintimes.co.uk/}
}
\usage{
trainstations
}
\description{
A reference dataset listing the codes, names and locations
of trains stations in Great Britain.
}
\keyword{datasets}
