% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{typeSafeIntersect}
\alias{typeSafeIntersect}
\title{Intersect two vectors without changing their data types.}
\usage{
typeSafeIntersect(x, y, dedupe = FALSE)
}
\arguments{
\item{x}{First set of values.}

\item{y}{Second set of values.}

\item{dedupe}{Specify TRUE to remove duplicate values.}
}
\value{
A vector containing the intersection of x and y.
}
\description{
\code{typeSafeIntersect} preserves data types in a way that the
\code{base::intersect} function does not, e.g. for Date values.
}
