% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{typeSafeUnion}
\alias{typeSafeUnion}
\title{Union two vectors without changing their data types.}
\usage{
typeSafeUnion(x, y, dedupe = FALSE)
}
\arguments{
\item{x}{First set of values.}

\item{y}{Second set of values.}

\item{dedupe}{Specify TRUE to remove duplicate values.}
}
\value{
A vector containing the union of x and y
}
\description{
\code{typeSafeUnion} preserves data types in a way that the
\code{base::union} function does not, e.g. for Date values.
}
