% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{typeSafeUnlist}
\alias{typeSafeUnlist}
\title{Unlist a list into a vector in a type-safe way where possible.}
\usage{
typeSafeUnlist(x)
}
\arguments{
\item{x}{A list to convert to a vector.}
}
\value{
A vector containing the values from x.
}
\description{
\code{typeSafeUnlist} tries to preserve data types in a way that the
\code{base::unlist} function does not for Date, POSIXct and POSIXlt values.
}
\details{
If a list containing mixed types is specified, then \code{typeSafeUnlist}
falls back to using \code{base::unlist}.
}
