% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueRangeExpression.R
\name{vreConvertSimpleNumericRange}
\alias{vreConvertSimpleNumericRange}
\title{Convert a simple range expression to a standard R logical expression.}
\usage{
vreConvertSimpleNumericRange(vre)
}
\arguments{
\item{vre}{The value range expression to examine.}
}
\value{
A standard R logical expression.
}
\description{
\code{vreConvertSimpleNumericRange} is a utility function that converts
a simple range expression of the form "value1<=v<value2" to a standard
R logical expression of the form "value1<=v && v<value2".
}
