% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueRangeExpression.R
\name{vreGetSingleValue}
\alias{vreGetSingleValue}
\title{Read the value from a single-valued value range expression.}
\usage{
vreGetSingleValue(vre)
}
\arguments{
\item{vre}{The value range expression to examine.}
}
\value{
The value read from the expression.
}
\description{
\code{vreGetSingleValue} is a utility function reads the single value
from a value range expression (it assumes the specified
is either numeric, a number expressed as
text or an expression of the form "v=" or "v==").
}
