% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueRangeExpression.R
\name{vreHexToClr}
\alias{vreHexToClr}
\title{Convert a colour in hex format (#RRGGBB) into a list.}
\usage{
vreHexToClr(hexclr)
}
\arguments{
\item{hexclr}{The colour to convert.}
}
\value{
The converted colour.
}
\description{
\code{vreHexToClr} converts a colour in hex format
(#RRGGBB) into a list of three element (r, g and b).
}
