% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueRangeExpression.R
\name{vreIsEqual}
\alias{vreIsEqual}
\title{Test if two numeric values are equal within tolerance.}
\usage{
vreIsEqual(value1, value2)
}
\arguments{
\item{value1}{The first value to compare.}

\item{value2}{The second value to compare.}
}
\value{
`TRUE` if the two numbers are equal, `FALSE` otherwise.
}
\description{
\code{vreIsEqual} tests whether two values are equal
within sqrt(.Machine$double.eps).
}
