% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueRangeExpression.R
\name{vreIsMatch}
\alias{vreIsMatch}
\title{Test whether a value matches a value range expression.}
\usage{
vreIsMatch(vre, v, testOnly = FALSE)
}
\arguments{
\item{vre}{The value range expression.}

\item{v}{The value.}

\item{testOnly}{`TRUE` if this comparison is just a test.}
}
\value{
`TRUE` if v matches the criteria specified in the value range expression,
`FALSE` otherwise.
}
\description{
\code{vreIsMatch} tests a value (e.g. from a cell) matches
the criteria specified in a value range expression.
}
