% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueRangeExpression.R
\name{vreIsSimpleNumericRange}
\alias{vreIsSimpleNumericRange}
\title{Determine if a value range expression is a simple range expression.}
\usage{
vreIsSimpleNumericRange(vre)
}
\arguments{
\item{vre}{The value range expression to examine.}
}
\value{
`TRUE` if vre is a simple range expression, `FALSE` otherwise.
}
\description{
\code{vreIsSingleValue} is a utility function that returns `TRUE` if the
specified value range expression is a simple range expression of the
form "value1<=v<value2", where the logical comparisons can be < or <= only
and the values must be numbers.
}
