% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueRangeExpression.R
\name{vreIsSingleValue}
\alias{vreIsSingleValue}
\title{Determine if a value range expression is a single value.}
\usage{
vreIsSingleValue(vre)
}
\arguments{
\item{vre}{The value range expression to examine.}
}
\value{
`TRUE` if vre is a single value, `FALSE` otherwise.
}
\description{
\code{vreIsSingleValue} is a utility function that returns `TRUE` if the
specified value range expression is either numeric, a number expressed as
text or an expression of the form "v=" or "v==".
}
