% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueRangeExpression.R
\name{vreScale2Colours}
\alias{vreScale2Colours}
\title{Scale a number from a range into a colour gradient.}
\usage{
vreScale2Colours(clr1, clr2, vMin, vMax, value)
}
\arguments{
\item{clr1}{The colour representing the lower value of the target range.}

\item{clr2}{The colour representing the upper value of the target range.}

\item{vMin}{The lower value of the source range.}

\item{vMax}{The upper value of the source range.}

\item{value}{The source value to rescale into the target range.}
}
\value{
The value scaled into the target colour gradient.
}
\description{
\code{vreScale2Colours} takes a value from a range and
scales it proportionally into a colour from a colour gradient.
}
