% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueRangeExpression.R
\name{vreScaleNumber}
\alias{vreScaleNumber}
\title{Rescale a number from one range into another range.}
\usage{
vreScaleNumber(n1, n2, vMin, vMax, value, decimalPlaces = 3)
}
\arguments{
\item{n1}{The lower value of the target range.}

\item{n2}{The upper value of the target range.}

\item{vMin}{The lower value of the source range.}

\item{vMax}{The upper value of the source range.}

\item{value}{The source value to rescale into the target range.}

\item{decimalPlaces}{The number of decimal places to round the result to.}
}
\value{
The value rescaled into the target range.
}
\description{
\code{vreScaleNumber} takes a value from one range and
scales it proportionally into another range.
}
