% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertor.R
\name{convertPvtTblToBasicTbl}
\alias{convertPvtTblToBasicTbl}
\title{Convert a pivot table to a basic table.}
\usage{
convertPvtTblToBasicTbl(
  pvt = NULL,
  exportOptions = NULL,
  compatibility = NULL,
  showRowGroupHeaders = FALSE
)
}
\arguments{
\item{pvt}{The pivot table to convert.}

\item{exportOptions}{Options specifying how values are exported.}

\item{compatibility}{Compatibility options specified when creating the
basictabler table.}

\item{showRowGroupHeaders}{Show captions at the top of the columns that
comprise the row groups (i.e. in the top left root of then pivot table).}
}
\value{
a basictabler table.
}
\description{
\code{convertPvtTblToBasicTbl} is a utility function that converts a pivot
table to a basic table from the basictabler package.
}
