% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotThemes.R
\name{getBlankTheme}
\alias{getBlankTheme}
\title{Get an empty theme for applying no styling to a table.}
\usage{
getBlankTheme(parentPivot, themeName = "blank")
}
\arguments{
\item{parentPivot}{Owning pivot table.}

\item{themeName}{The name to use as the new theme name.}
}
\value{
A TableStyles object.
}
\description{
Get an empty theme for applying no styling to a table.
}
