% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PivotBatchCalculator.R
\docType{class}
\name{PivotBatchCalculator}
\alias{PivotBatchCalculator}
\title{A class that calculates the values for multiple cells.}
\format{\code{\link{R6Class}} object.}
\usage{
PivotBatchCalculator
}
\value{
Object of \code{\link{R6Class}} with properties and methods that help
  to perform calculations in batch.
}
\description{
The PivotBatchCalculator class calculates the values for multiple cells in
the pivot table in one evaluation step.
}
\section{Fields}{

\describe{
\item{\code{parentPivot}}{Owning pivot table.}
}}

\section{Methods}{

\describe{
  \item{Documentation}{For more complete explanations and examples please see
  the extensive vignettes supplied with this package.}
  \item{\code{new(...)}}{Create a new batch calculator.}

  \item{\code{exampleMethod(df, dataName)}}{Example to be removed later.}
}
}

\examples{
# This class should only be created by the pivot table.
# It is not intended to be created outside of the pivot table.
}
\keyword{datasets}
