% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivottabler.R
\name{renderPivottabler}
\alias{renderPivottabler}
\title{Standard function for Shiny scaffolding.}
\usage{
renderPivottabler(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{The R expression to execute and render in the Shiny web application.}

\item{env}{Standard shiny argument for a render function.}

\item{quoted}{Standard shiny argument for a render function.}
}
\description{
Standard function for Shiny scaffolding.
}
